CovModel: add 'set_opt_arg' routine to tools to shrink __init__
